+++ netbox_pdu_plugin/filtersets.py
@@ -16,7 +16,7 @@ class ManagedPDUFilterSet(NetBoxModelFilterSet):
 
     class Meta:
         model = ManagedPDU
-        fields = ("id", "device", "sync_status")
+        fields = ("id", "device", "sync_status", "sync_enabled", "metrics_enabled")
 
     def search(self, queryset, name, value):
         return queryset.filter(device__name__icontains=value)